fix(deepseek-v4): fp32 grad-touch + lint wrap
- train_ft.py: promote logits to fp32 before the zero-valued full-logits grad
  touch. Low-precision (fp16) logits over a large vocab can overflow to inf, and
  inf * 0.0 = nan would poison local_loss and the backward pass. (Review catch.)
- test_train_ft.py: emit wide large-magnitude low-precision logits and assert the
  loss/grads stay finite, guarding the fp32 promotion.
- layers.py: wrap two over-length _query_positions calls (ruff format / CI lint).

+++ nemo_automodel/recipes/llm/train_ft.py
@@ -1272,7 +1272,10 @@ class TrainFinetuneRecipeForNextTokenPrediction(BaseRecipe):
                 if is_train and batch.get("_cp_full_logits_grad_touch"):
                     logits = getattr(out, "logits", out)
                     if isinstance(logits, torch.Tensor):
-                        local_loss = local_loss + logits.sum() * 0.0
+                        # Promote to fp32 before summing: bf16 logits over a large
+                        # vocab (e.g. DSV4's 129280) overflow to inf, and inf * 0.0
+                        # would be nan, poisoning local_loss and the backward pass.
+                        local_loss = local_loss + logits.float().sum() * 0.0
                 loss_buffer.append(local_loss.clone().detach())
                 if is_train:
                     (local_loss * self._get_dp_group_size(include_cp=True)).backward()
